# week-6/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Background Jobs API",
    "version": "1.0.0",
    "description": "Accept slow work (an AI summarization call) with 202 Accepted, run it in a background worker with retries, idempotency and alerting, and poll for the result."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API info",
        "responses": {
          "200": {
            "description": "Returns API metadata",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Server is alive",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "List jobs",
        "description": "Optionally filter by status: queued, running, completed, failed",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["queued", "running", "completed", "failed"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Job" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit a job (slow AI call runs in the background)",
        "description": "Responds instantly with 202 Accepted and a job id. Poll GET /jobs/{id} for the result. Set the Idempotency-Key header (or idempotency_key in the body) to guarantee the same job is not enqueued twice.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "Replaying the same key returns the existing job instead of creating a new one"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/JobRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted — job queued (or existing job returned for a replayed idempotency key)",
            "headers": {
              "Location": { "schema": { "type": "string" }, "description": "URL to poll for job status" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Job" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "summary": "Get job status and result",
        "description": "status: queued -> running -> completed | failed. attempts counts every run (including retries).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job with current status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Job" }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "summary": "List alerts",
        "description": "Alerts are raised when a job exhausts its retries. Also written to stderr and optionally POSTed to ALERT_WEBHOOK_URL.",
        "responses": {
          "200": {
            "description": "Array of alerts (newest first)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Alert" }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "JobRequest": {
        "type": "object",
        "required": ["type", "payload"],
        "properties": {
          "type": { "type": "string", "example": "summarize" },
          "idempotency_key": { "type": "string", "description": "Alternative to the Idempotency-Key header" },
          "payload": {
            "type": "object",
            "required": ["prompt"],
            "properties": {
              "prompt": { "type": "string", "example": "Summarize the plot of Dune in one paragraph." },
              "model": { "type": "string", "example": "gpt-4o-mini", "description": "Ignored by the offline mock; used when OPENAI_API_KEY is set" },
              "failOnce": { "type": "boolean", "description": "Demo retries: fail the first attempt, succeed on retry" },
              "failAlways": { "type": "boolean", "description": "Demo alerts: always fail until max_attempts is exhausted" }
            }
          }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "payload": { "type": "object" },
          "idempotency_key": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["queued", "running", "completed", "failed"] },
          "attempts": { "type": "integer" },
          "max_attempts": { "type": "integer" },
          "progress": { "type": "integer", "description": "0-100 while running" },
          "result": { "type": "object", "nullable": true, "description": "Set when completed" },
          "error": { "type": "string", "nullable": true },
          "created_at": { "type": "string" },
          "started_at": { "type": "string", "nullable": true },
          "finished_at": { "type": "string", "nullable": true },
          "next_run_at": { "type": "string", "nullable": true, "description": "When a retry is due (backoff)" }
        }
      },
      "Alert": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "job_id": { "type": "string" },
          "message": { "type": "string" },
          "created_at": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}